requirements.txt
Flask
gunicorn
python-telegram-bot==20.0
requests
